added generator file, run login before inquire
