+++ lib/firebase.js
@@ -51,7 +51,7 @@ module.exports = {
     }
     // assign process env FIREBASE_TOKEN to token
     process.env.FIREBASE_TOKEN = token;
-    this.useAdd(projectDirName, firebaseName);
+    // this.useAdd(projectDirName, firebaseName);
   },
 
   useAdd(projectDirName, firebaseName) {
