+++ lib/inquirer.js
@@ -5,7 +5,7 @@ const commands = require('./commands');
 const firebase = require('./firebase');
 
 module.exports = {
-  generateTemplate: async () => {
+  askTemplate: async () => {
     // Grab templates and filter out hidden files
     const CHOICES = fs.readdirSync(`${__dirname}/templates`).filter(file => file[0] !== '.');
     const QUESTIONS = [
@@ -30,38 +30,8 @@ module.exports = {
         },
       },
     ];
-    const CURR_DIR = process.cwd();
-
-    function createDirectoryContents(templatePath, newProjectPath) {
-      const filesToCreate = fs.readdirSync(templatePath);
-      filesToCreate.forEach((file) => {
-        const origFilePath = `${templatePath}/${file}`;
-        // get stats about the current file
-        const stats = fs.statSync(origFilePath);
-        if (stats.isFile()) {
-          const contents = fs.readFileSync(origFilePath, 'utf8');
-          const writePath = `${CURR_DIR}/${newProjectPath}/${file}`;
-          fs.writeFileSync(writePath, contents, 'utf8');
-        } else if (stats.isDirectory()) {
-          fs.mkdirSync(`${CURR_DIR}/${newProjectPath}/${file}`);
-          // recursive call
-          createDirectoryContents(`${templatePath}/${file}`, `${newProjectPath}/${file}`);
-        }
-      });
-    }
-    inquirer.prompt(QUESTIONS)
-      .then((answers) => {
-        const firebaseName = answers['firebase-name'];
-        const projectChoice = answers['project-choice'];
-        const projectName = answers['project-name'];
-        const templatePath = `${__dirname}/templates/${projectChoice}`;
-        fs.mkdirSync(`${CURR_DIR}/${projectName}`);
-        createDirectoryContents(templatePath, projectName);
-        console.log(chalk.blue(`Finished building ${projectName}`));
-        commands.changeDir(projectName);
-        // firebase.installFB();
-        firebase.FBLogin(projectName, firebaseName);
-      });
+    
+    return inquirer.prompt(QUESTIONS)
   },
   redeploy: async () => {
     // Grab templates and filter out hidden files
